Player HUD working now

+++ Assets/Plugins/GameCreator/Melee/Mono/Components/UI/CharacterMeleeUI.cs
@@ -42,7 +42,7 @@
 
         private void UpdateUI()
         {
-            healthSlider.value = melee.GetHP() / melee.maxHealth;
+            healthSlider.value = melee.GetHP() / (float)melee.maxHealth;
             if (melee.currentShield) defenseSlider.value = melee.Defense / melee.currentShield.maxDefense.GetValue(gameObject);
             poiseSlider.value = melee.Poise / melee.maxPoise.GetValue(gameObject);
         }
